msm8976-common: Change the PowerSaveOffload to 5
Default value 2 is more agressive for power save.
However we saw some degragation for voice quality.
Instead, 5 is used for VoIP call solution.
When the parameter is set to 5 and voice packets are detected
in uplink or downlink data
traffic, the QPower inactivity timeout is reset to the default
QPower-disabled value (200 ms).
This means that QPower power saving is not be triggered until after
200ms of inactivity

Change-Id: I1cd8ae9f8b39a998795f061cf41d72f68f694c41